Rewrite github-rs to use async Hyper and more
This is fundamentally a complete rewrite of github-rs. One of the most major
parts is that this uses the long awaited asynchronous version of Hyper that
leverages tokio and futures. While this isn't release ready yet since hyper 0.11
has not been released I find that this has been a major update, enough so that
it's in a stable enough place to continue growing as a library and that this is
significant turning point for it. Here's what's changed and acts as the basis
for the upcoming 0.5 release:

  - github-rs the async version of Hyper under the hood
  - It uses a builder pattern now rather than individual function calls
  - The underlying code is mostly generated by macros since I've settled on
    a good design.
  - Lots of docs have been added for the project, significantly more polished
    than before (of course there's always room for improvement)
  - The code is better separated and more logically structured.
  - Deserialization to specific structs has been removed and favors using
    Serde's generic Value type as well as returning the status code. This favors
    GitHub's API design on the whole and allows focusing on the endpoints
    themselves and not what comes out of it.
  - Switching to a dual license for all code which is what I use for most of my
    projects now.

These were the major points and while there are smaller improvements I've found
that these were the main focus of the rewrite and have laid a strong foundation
for further updates down the line.