Update macros and macro syntax
This commit rewrites many of the core macros. This allows for easier syntax to
read for deriving From traits for types as well as generating impls for the
code. All the parts affected by the change have been rewritten but this doesn't
affect functionality at all. It just makes it easier for the library maintainers