Make the library async
Currently, the library is using tokio/futures internally, but exposes an
entirely synchronous interface, thus losing most of the advantages of
futures/tokio:

- Running requests in parallel.
- Racing requests.
- Chaining requests.
- Applying generic timeouts.
- Sharing the event loop (e.g. performing an asynchronous request using
  this API from an async HTTP server, without spawning a thread).
- Etc. etc.

A synchronous interface requires using threads for such things, and its
implementation itself would be much simpler just using threads and
blocking IO, instead of the complications of event loops.

Instead, this:

- Exposes futures/tokio as a public dependency.
- Take a tokio_core &Handle instead of maintaining own internal Core.
- Returns a Future from .execute().

The internals are simplified, and the API remains almost the same, just
needing to add a few `core.run(...)` calls, as can be seen in the
examples.