macros: introduce executor trait
This changes the macro structure so that instead of using from! and
impl_macro! to create the execute() method, it's done by invoking the
exec! macro. This allows all types with the execute() method to
implement the Executor trait. By implementing a trait, it allows
consuming code to easily pass queries around without having to execute
them up front.