Allow stricter type checking to API GET/POST parameters
Some GET/POST parameters only admit a small set of enumerated values.
For example, only "forks", "stars", and "updated" are allowed for the
"sort" GET parameter in a /search/repositories query.

This commit tweaks the |?> form of the impl_macro!() macro to specify a
type for the input to the generated function. For unrestricted
parameters, this will continue to be &str. For parameters with a
restricted set of values, this can now be an enum. The enum's
to_string() implementation is used to supply the value of the parameter
in the submitted query.

This commit adds also appropriate enums for all of the non-preview
search endpoint parameters.