testutil: Convert DOS line endings to Unix
This commit replaces the CRLF line ending with just LF.

The Git client has the core.autocrlf feature, which allows you to
automatically convert from the standard in git repository storage (LF)
to one's favorite LF style.  Most of the other files in this
repository use LF endings, so this commit brings this file in parity
with the rest of the project.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>